--- taskchampion/tests/shared/mod.rs
@@ -1,3 +1,0 @@
-mod testserver;
-
-pub use testserver::TestServer;
